third_party/WebKit/LayoutTests/html/dialog/inert-inlines-expected.txt

Tests that inert inlines do not receive mouse events. To test manually, click on all the "Click me"s. The test fails if you see red.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


clicking on a
PASS inertElementFiredOn is false
PASS expectedTarget.firedOn is true
clicking on button
PASS inertElementFiredOn is false
PASS expectedTarget.firedOn is true
clicking on div
PASS inertElementFiredOn is false
PASS expectedTarget.firedOn is true
clicking on span
PASS inertElementFiredOn is false
PASS expectedTarget.firedOn is true
PASS successfullyParsed is true

TEST COMPLETE
Click me Click me
Click meeee
Click me
Click meeee
